backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
